requirements.txt
pre-commit==3.7.1
python-telegram-bot==21.2
requests
httpx
bech32
python-dotenv
